// src/app/models/product.model.ts
export interface ProductModel {
  id: number;
  code: string;
  name: string;
  category: string;
  supplier: string;
  brand: string;
  model: string;
  color: string;
  country: string;
  price: number;
  cost: number;
  quantity: number;
  reserved: number;
  sold: number;
  rating: number;
  weight: number;
  width: number;
  height: number;
  depth: number;
  warranty: number;
  createdDate: string;
  updatedDate: string;
  isActive: boolean;
  status: string;
  description: string;
}